requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
python-decouple==3.8
requests==2.28.2
pytz==2023.3.post1